base class for the error generation and integration handling
